+++ src/Link/ApiLinkMacro.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Tomaj\NetteApi\Link;
 
 use Latte\Compiler;