+++ src/Validation/JsonSchemaValidator.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Tomaj\NetteApi\Validation;
 
 use JsonSchema\Validator;