+++ tests/Presenters/ApiPresenterTest.php
@@ -17,7 +17,7 @@ use Tomaj\NetteApi\EndpointIdentifier;
 use Tomaj\NetteApi\Handlers\AlwaysOkHandler;
 use Tomaj\NetteApi\Handlers\EchoHandler;
 use Tomaj\NetteApi\Misc\IpDetector;
-use Tomaj\NetteApi\Misc\StaticBearerTokenRepository;
+use Tomaj\NetteApi\Misc\StaticTokenRepository;
 use Tomaj\NetteApi\Presenters\ApiPresenter;
 use Tomaj\NetteApi\Test\Handler\TestHandler;
 use Tracy\Debugger;
@@ -49,7 +49,7 @@ class ApiPresenterTest extends TestCase
         $apiDecider->addApi(
             new EndpointIdentifier('GET', 1, 'test', 'api'),
             new AlwaysOkHandler(),
-            new BearerTokenAuthorization(new StaticBearerTokenRepository([]), new IpDetector())
+            new BearerTokenAuthorization(new StaticTokenRepository([]), new IpDetector())
         );
 
         $presenter = new ApiPresenter();
@@ -78,17 +78,19 @@ class ApiPresenterTest extends TestCase
         $presenter->response = new HttpResponse();
         $presenter->context = new Container();
 
+        Debugger::$productionMode = Debugger::PRODUCTION;
+
         $request = new Request('Api:Api:default', 'GET', ['version' => 1, 'package' => 'test', 'apiAction' => 'api']);
         $result = $presenter->run($request);
 
         $this->assertEquals(['status' => 'error', 'message' => 'wrong input'], $result->getPayload());
         $this->assertEquals('application/json', $result->getContentType());
 
-        Debugger::enable(true);
+        Debugger::$productionMode = Debugger::DETECT;
+
         $result = $presenter->run($request);
         $this->assertEquals(['status' => 'error', 'message' => 'wrong input', 'detail' => ['status' => ['Field is required']]], $result->getPayload());
         $this->assertEquals('application/json', $result->getContentType());
-        Debugger::enable(false);
     }
 
     public function testWithOutputs()